2fa code: improve speed by reimplementing code gen code
