+++ steamctl/commands/authenticator/__init__.py
@@ -52,7 +52,7 @@ def cmd_parser(cp):
 
     scp = sub_cp.add_parser("code", help="Generate auth code")
     scp.add_argument('account', type=str, help='Account name').completer = account_autocomplete
-    scp.set_defaults(_cmd_func=__name__ + '.cmds:cmd_authenticator_code')
+    scp.set_defaults(_cmd_func=__name__ + '.cmd_code:cmd_authenticator_code')
 
     scp = sub_cp.add_parser("qrcode", help="Generate QR code")
     scp.add_argument('--compat', action='store_true',
@@ -64,5 +64,5 @@ def cmd_parser(cp):
                      help='Invert QR code colors. Try if app fails to scan the code.'
                      )
     scp.add_argument('account', type=str, help='Account name').completer = account_autocomplete
-    scp.set_defaults(_cmd_func=__name__ + '.cmds:cmd_authenticator_qrcode')
+    scp.set_defaults(_cmd_func=__name__ + '.cmd_code:cmd_authenticator_qrcode')
