+++ steamctl/commands/authenticator/cmds.py
@@ -346,56 +346,3 @@ def cmd_authenticator_status(args):
         print("Classified agent:", status['classified_agent'])
         print("Revocation attempts remaining:", status['revocation_attempts_remaining'])
 
-def cmd_authenticator_code(args):
-    account = args.account.lower().strip()
-    secrets = UserDataFile('authenticator/{}.json'.format(account)).read_json()
-
-    if not secrets:
-        print("No authenticator for %r" % account)
-        return 1  # error
-
-    print(SteamAuthenticator(secrets).get_code())
-
-def cmd_authenticator_qrcode(args):
-    account = args.account.lower().strip()
-    secrets = UserDataFile('authenticator/{}.json'.format(account)).read_json()
-
-    if not secrets:
-        print("No authenticator for %r" % account)
-        return 1  # error
-
-    import pyqrcode
-    from base64 import b64decode, b32encode
-
-    if args.invert:
-        FG, BG = '0', '1'
-    else:
-        FG, BG = '1', '0'
-
-    charmap = {
-      (BG, BG): '█',
-      (FG, FG): ' ',
-      (BG, FG): '▀',
-      (FG, BG): '▄',
-    }
-
-    if args.compat:
-        uri = 'otpauth://totp/steamctl:{user}?secret={secret}&issuer=Steam&digits=5'
-    else:
-        uri = 'otpauth://steam/steamctl:{user}?secret={secret}&issuer=Steam'
-
-
-    uri = uri.format(user=secrets['account_name'],
-                     secret=b32encode(b64decode(secrets['shared_secret'])).decode('ascii'),
-                     )
-
-    qrlines = pyqrcode.create(uri, error='M').text(1).split('\n')[:-1]
-
-    print("Suggested 2FA App: Aegis, andOTP")
-    print("Scan the QR code below:")
-
-    for y in range(0, len(qrlines), 2):
-        for x in range(0, len(qrlines[y])):
-            print(charmap[(qrlines[y][x], FG if y+1 >= len(qrlines) else qrlines[y+1][x])], end='')
-        print()
-
